feat: implement BlogPostView component for blog draft previews
