+++ pages/blog/[slug].vue
@@ -10,7 +10,6 @@
           <NuxtLink
             to="/blog/"
             class="blog__back"
-            :class="tableOfContents.length ? 'blog__back__margin' : ''"
             @click.prevent="goBack"
           >
             <span>← Back</span>
@@ -45,35 +44,6 @@
             </span>
           </div>
         </div>
-        <nav
-          v-if="tableOfContents.length"
-          class="navbar navbar-expand bg-white py-3"
-        >
-          <div class="collapse navbar-collapse">
-            <ul class="navbar-nav">
-              <li class="nav-item dropdown">
-                <a
-                  class="dropdown-toggle"
-                  href="#"
-                  id="tocMenuLink"
-                  role="button"
-                  data-bs-toggle="dropdown"
-                  aria-expanded="false"
-                >
-                  Table of Contents
-                </a>
-                <ul class="dropdown-menu" aria-labelledby="tocMenuLink">
-                  <li v-for="link of tableOfContents" :key="link.id">
-                    <NuxtLink class="dropdown-link" :to="`#${link.id}`">
-                      {{ link.text }}
-                    </NuxtLink>
-                  </li>
-                </ul>
-              </li>
-            </ul>
-          </div>
-        </nav>
-        <h1 class="mb-1 article__heading">{{ blogData?.title }}</h1>
         <div class="d-flex sm-text my-2 datentimeToRead">
           <span>{{ formatDate(blogData?.publishedAt) }}</span>
           <span>|</span>
@@ -84,26 +54,7 @@
             <span>{{ blogData?.readingStats?.text }}</span>
           </div>
         </div>
-        <div class="sm-text mt-1 article__author-section">
-          by
-          <a
-            :href="blogData?.authorProfile"
-            :title="blogData?.authorProfile"
-            target="_blank"
-            rel="noopener"
-          >
-            <span class="article__author">{{ blogData?.author }}</span>
-          </a>
-        </div>
-        <div class="blog__content" ref="blogContent">
-          <!-- <nuxt-content :document="processedBlogData" v-if="processedBlogData" /> -->
-          <div class="nuxt-content" v-html="processedBlogData"/>
-
-          <div class="popup__img" ref="popupImg" :style="{ display: showPopup ? 'block' : 'none' }">
-            <span class="image-preview-close" @click="closePopup">&times;</span>
-            <img :src="popupImageSrc" :alt="popupImageAlt" @click="closePopup" ref="popupImage" />
-          </div>
-        </div>
+        <BlogPostView v-if="blogPostViewData" :blog-data="blogPostViewData" />
         <notifications position="bottom right" class="my-notification" />
 
         <div v-if="relatedArticles" class="mt-5">
@@ -145,17 +96,11 @@ import CopyLinkIcon from '../../components/icons/copyLink.vue'
 import getSiteMeta from '../../utils/getSiteMeta'
 import readingTime from '@/utils/readingTime'
 import { getBlogBySlug, getAllBlogs } from '@/utils/getAllBlogs'
-import { marked } from 'marked'
 
 const route = useRoute()
 const router = useRouter()
 const config = useRuntimeConfig()
 const { $notify } = useNuxtApp()
-
-const showPopup = ref(false)
-const popupImageSrc = ref('')
-const popupImageAlt = ref('')
-const blogContent = ref(null)
 
 const { data: blogResponse } = await useAsyncData(`blog-${route.params.slug}`, async () => {
   try {
@@ -193,38 +138,17 @@ const { data: blogResponse } = await useAsyncData(`blog-${route.params.slug}`, a
 const blogData = computed(() => blogResponse.value?.blogData)
 const relatedArticles = computed(() => blogResponse.value?.relatedArticles || [])
 
-const setupImageClickHandlers = () => {
-  nextTick(() => {
-    if (blogContent.value) {
-      const images = blogContent.value.querySelectorAll('.nuxt-content img')
-      images.forEach((image) => {
-        image.addEventListener('click', () => {
-          openPopup(image.src, image.alt)
-        })
-      })
-    }
-  })
-}
-
-const setupKeyboardHandler = () => {
-  document.addEventListener('keydown', (evt) => {
-    if (evt.key === 'Escape') {
-      closePopup()
-    }
-  })
-}
-
-const openPopup = (src, alt) => {
-  popupImageSrc.value = src
-  popupImageAlt.value = alt
-  showPopup.value = true
-}
-
-const closePopup = () => {
-  showPopup.value = false
-  popupImageSrc.value = ''
-  popupImageAlt.value = ''
-}
+const blogPostViewData = computed(() => {
+  if (!blogData.value) return null
+  return {
+    title: blogData.value.title,
+    body: blogData.value.body,
+    author: blogData.value.author,
+    authorProfile: blogData.value.authorProfile,
+    coverImgUrl: null,
+    coverImgAlt: blogData.value.coverImgAlt,
+  }
+})
 
 const formatDate = (date) => {
   const options = { year: 'numeric', month: 'long', day: 'numeric' }
@@ -273,83 +197,6 @@ const meta = computed(() => {
 
 const encodedUrl = computed(() => {
   return encodeURIComponent(config.public.baseUrl + route.fullPath)
-})
-
-const processedBlogData = computed(() => {
-  const renderer = new marked.Renderer()
-
-  renderer.heading = function(text, level) {
-    if (level >= 2 && level <= 3) {
-      const id = text
-        .toLowerCase()
-        .replace(/[^\w\s-]/g, '')
-        .replace(/\s+/g, '-')
-        .replace(/--+/g, '-')
-        .replace(/^-+|-+$/g, '')
-
-      return `<h${level} id="${id}">${text}</h${level}>`
-    }
-    return `<h${level}>${text}</h${level}>`
-  }
-
-  renderer.link = function(href, title, text) {
-    // Only add trailing slash to internal links
-    let processedHref = href
-
-    // Check if it's an internal link (relative or formester.com)
-    const isInternal = href.startsWith('/') || href.includes('formester.com')
-
-    if (isInternal && href) {
-      // Don't add trailing slash if:
-      // - Already has trailing slash
-      // - Has file extension (like .pdf, .jpg, etc.)
-      // - Has hash fragment or query string
-      const hasTrailingSlash = href.endsWith('/')
-      const hasFileExtension = /\.[a-z0-9]+$/i.test(href.split(/[?#]/)[0])
-      const hasHashOrQuery = href.includes('#') || href.includes('?')
-
-      if (!hasTrailingSlash && !hasFileExtension && !hasHashOrQuery) {
-        processedHref = href + '/'
-      }
-    }
-
-    const titleAttr = title ? ` title="${title}"` : ''
-    return `<a href="${processedHref}"${titleAttr}>${text}</a>`
-  }
-
-  return marked(blogData.value?.body || '', { renderer })
-})
-
-const tableOfContents = computed(() => {
-  const toc = []
-  const content = blogData.value?.body || ''
-
-  const headingRegex = /^(#{2,3})\s+(.+)$/gm
-  let match
-
-  while ((match = headingRegex.exec(content)) !== null) {
-    const level = match[1].length
-    const text = match[2].trim()
-    const id = text
-      .toLowerCase()
-      .replace(/[^\w\s-]/g, '')
-      .replace(/\s+/g, '-')
-      .replace(/--+/g, '-')
-      .replace(/^-+|-+$/g, '')
-
-    toc.push({
-      id,
-      text,
-      level
-    })
-  }
-
-  return toc
-})
-
-onMounted(() => {
-  setupImageClickHandlers()
-  setupKeyboardHandler()
 })
 
 useHead({
@@ -455,19 +302,9 @@ useJsonld(jsonldData.value)
 p {
   margin-bottom: 2rem;
 }
-nav {
-  position: sticky;
-  top: 72px;
-  z-index: 100;
-}
-.article-container{
-    margin-top: 9rem;
-}
-.article__heading {
-  font-size: 2.25rem;
-  font-weight: 700;
-  line-height: 44px;
-  color: var(--clr-text-primary);
+
+.article-container {
+  margin-top: 9rem;
 }
 
 .article__sub-heading {
@@ -483,15 +320,10 @@ nav {
   color: var(--clr-text-primary);
 }
 
-
 .sm-text {
   font-size: 14px;
   line-height: 21px;
   color: hsla(0, 0%, 31%, 1);
-}
-
-.article__author {
-  font-weight: 600;
 }
 
 .mw-920 {
@@ -506,10 +338,6 @@ nav {
   margin-bottom: 3rem;
 }
 
-.article__author-section {
-  opacity: 0.75;
-}
-
 .datentimeToRead {
   gap: 0.75rem;
   opacity: 0.5;
@@ -519,193 +347,24 @@ nav {
   gap: 0.5rem;
 }
 
-#tocMenuLink {
-  color: #777;
-  font-size: 16px;
-}
-
-.dropdown-menu {
-  border: 1px solid #e5e7eb;
-  border-radius: 8px;
-  box-shadow: 0 10px 25px -3px rgba(0, 0, 0, 0.1), 0 4px 6px -2px rgba(0, 0, 0, 0.05);
-  backdrop-filter: blur(12px);
-  background-color: rgba(255, 255, 255, 0.95);
-  max-height: 450px;
-  overflow-y: auto;
-  overflow-x: hidden;
-  padding: 8px 0;
-  margin-top: 8px;
-}
-
-.dropdown-link {
-  padding: 12px 16px;
-  min-width: 750px;
-  width: 100%;
-  display: block;
-  color: #374151;
-  text-decoration: none;
-  font-size: 14px;
-  line-height: 1.4;
-  transition: all 0.2s ease;
-  border-left: 3px solid transparent;
-}
-
-.dropdown-link:hover {
-  background-color: #f3f4f6;
-  color: var(--clr-primary);
-  border-left-color: var(--clr-primary);
-  text-decoration: none;
-}
-
 .blog__header {
   margin-top: -4rem;
   justify-content: space-between;
   display: flex;
 }
+
 .blog__header span {
   color: #777;
 }
+
 .social__links {
   display: flex;
 }
+
 .social-icons {
   margin: 0 6px;
   cursor: pointer;
-}
-.social-icons {
   fill: #000;
 }
-
-.blog__content img {
-  cursor: zoom-in;
-}
-
-.popup__img {
-  position: fixed;
-  top: 0;
-  left: 0;
-  background: rgba(90, 90, 90, 0.96);
-  height: 100%;
-  width: 100%;
-  z-index: 99999;
-  display: none;
-}
-
-.popup__img span {
-  position: absolute;
-  top: 0;
-  right: 10px;
-  font-size: 30px;
-  font-weight: bolder;
-  color: #fff;
-  cursor: pointer;
-  z-index: 1;
-}
-
-.popup__img img {
-  cursor: zoom-out;
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  width: 80%;
-  object-fit: cover;
-}
-
-@media only screen and (max-width: 992px) {
-  nav {
-    top: 60px;
-  }
-  .dropdown-link {
-    min-width: 680px;
-  }
-}
-@media only screen and (max-width: 768px) {
-  .dropdown-link {
-    min-width: 480px;
-  }
-
-  .nuxt-content h2 {
-    margin-top: 16px;
-  }
-
-  .nuxt-content h3 {
-    margin-top: 16px;
-  }
-
-  .nuxt-content img {
-    margin-top: 8px;
-    margin-bottom: 12px;
-  }
-
-  .popup__img img {
-    width: 90%;
-  }
-}
-
-@media only screen and (max-width: 576px) {
-  .dropdown-link {
-    min-width: 370px;
-  }
-}
-@media only screen and (max-width: 432px) {
-  .dropdown-link {
-    min-width: 320px;
-  }
-}
-@media only screen and (max-width: 360px) {
-  .dropdown-link {
-    min-width: 250px;
-  }
-}
 </style>
 
-<style>
-.nuxt-content {
-  margin-top: 18px;
-  font-size: 16px;
-  line-height: 24px;
-  letter-spacing: 0.4px;
-  color: var(--clr-text-primary);
-}
-
-.nuxt-content p {
-  margin-bottom: 1.4rem;
-}
-
-.nuxt-content h2 {
-  font-size: 30px;
-  font-weight: 700;
-  line-height: 38px;
-  color: var(--clr-text-primary);
-  margin-top: 64px;
-  margin-bottom: 24px;
-}
-
-.nuxt-content h3 {
-  font-size: 24px;
-  font-weight: 600;
-  line-height: 28px;
-  color: var(--clr-text-primary);
-  margin-top: 48px;
-  margin-bottom: 24px;
-}
-
-.nuxt-content a {
-  color: var(--clr-primary);
-  text-decoration: underline;
-}
-
-.nuxt-content img {
-  height: auto;
-  width: 100%;
-  margin-top: 8px;
-  margin-bottom: 16px;
-}
-
-.nuxt-content ul li,
-.nuxt-content ol li {
-  margin-top: 0px;
-  margin-bottom: 16px;
-}
-</style>
